f2

+++ src/com/yatsenko/java1101/Homework1/Homework4/Math1.java
@@ -6,7 +6,7 @@ import static java.lang.Math.sin;
 
 public class Math1 {
     public static void main (String []args){
-        int a = 30;
+        double a = 30;
         double rad;
 
             rad = Math.toRadians(a);
